Improve self-update feature
