+++ cli/cli.go
@@ -20,9 +20,6 @@ import (
 	"github.com/essentialkaos/ek/v13/fsutil"
 	"github.com/essentialkaos/ek/v13/options"
 	"github.com/essentialkaos/ek/v13/req"
-	"github.com/essentialkaos/ek/v13/selfupdate"
-	"github.com/essentialkaos/ek/v13/selfupdate/interactive"
-	"github.com/essentialkaos/ek/v13/selfupdate/storage/basic"
 	"github.com/essentialkaos/ek/v13/strutil"
 	"github.com/essentialkaos/ek/v13/support"
 	"github.com/essentialkaos/ek/v13/support/deps"
@@ -150,7 +147,7 @@ func Run(gitRev string, gomod []byte) {
 			WithResources(resources.Collect()).
 			Print()
 		os.Exit(0)
-	case options.GetB(OPT_UPDATE):
+	case withSelfUpdate && options.GetB(OPT_UPDATE):
 		os.Exit(updateBinary())
 	case options.GetB(OPT_HELP) || len(args) == 0:
 		genUsage().Print()
@@ -407,39 +404,6 @@ func printErrorAndExit(f string, a ...interface{}) {
 
 // ////////////////////////////////////////////////////////////////////////////////// //
 
-// updateBinary updates current binary to the latest version
-func updateBinary() int {
-	quiet := strings.ToLower(options.GetS(OPT_UPDATE)) == "quiet"
-	updInfo, hasUpdate, err := basic.NewStorage("https://apps.kaos.ws").Check(APP, VER)
-
-	if err != nil {
-		if !quiet {
-			terminal.Error("Can't update binary: %v", err)
-		}
-
-		return 1
-	}
-
-	if !hasUpdate {
-		fmtc.If(!quiet).Println("{g}You are using the latest version of the app{!}")
-		return 0
-	}
-
-	pubKey := "MFkwEwYHKoZIzj0CAQYIKoZIzj0DAQcDQgAEnYHsOTvrKqeE97dsEt7Ge97+yUcvQJn1++s++FqShDyqwV8CcoKp0E6nDTc8SxInZ5wxwcScxSicfvC9S73OSg=="
-
-	if quiet {
-		err = selfupdate.Run(updInfo, pubKey, nil)
-	} else {
-		err = selfupdate.Run(updInfo, pubKey, interactive.Dispatcher())
-	}
-
-	if err != nil {
-		return 1
-	}
-
-	return 0
-}
-
 // printCompletion prints completion for given shell
 func printCompletion() int {
 	info := genUsage()
@@ -485,8 +449,12 @@ func genUsage() *usage.Info {
 	info.AddOption(OPT_QUIET, "Quiet mode")
 	info.AddOption(OPT_IGNORE_PACKAGES, "Do not check system for installed packages")
 	info.AddOption(OPT_NO_CLEANUP, "Disable deleting files created during tests")
+
+	if withSelfUpdate {
+		info.AddOption(OPT_UPDATE, "Update application to the latest version")
+	}
+
 	info.AddOption(OPT_NO_COLOR, "Disable colors in output")
-	info.AddOption(OPT_UPDATE, "Update application to the latest version")
 	info.AddOption(OPT_HELP, "Show this help message")
 	info.AddOption(OPT_VER, "Show version")
