+++ cli/without_selfupdate.go
@@ -1,0 +1,22 @@
+//go:build !selfupdate
+// +build !selfupdate
+
+package cli
+
+// ////////////////////////////////////////////////////////////////////////////////// //
+//                                                                                    //
+//                         Copyright (c) 2025 ESSENTIAL KAOS                          //
+//      Apache License, Version 2.0 <https://www.apache.org/licenses/LICENSE-2.0>     //
+//                                                                                    //
+// ////////////////////////////////////////////////////////////////////////////////// //
+
+var withSelfUpdate = false
+
+// ////////////////////////////////////////////////////////////////////////////////// //
+
+// updateBinary updates current binary to the latest version
+func updateBinary() int {
+	return 1
+}
+
+// ////////////////////////////////////////////////////////////////////////////////// //
